refactor: API 요청 로직 분리
