+++ db.json
@@ -35,12 +35,6 @@
       "name": "도스타코스 선릉점",
       "description": "멕시칸 캐주얼 그릴",
       "category": "기타"
-    },
-    {
-      "id": "하오츠",
-      "category": "중식",
-      "name": "하오츠",
-      "description": "ㅇ"
     }
   ]
 }
